requirements.txt
pandas
numpy
python-dotenv
requests